move get_user_by_id to users and authService
