+++ src/interface/src/app/services/plan.service.spec.ts
@@ -94,28 +94,6 @@ describe('PlanService', () => {
     });
   });
 
-  describe('getUser', () => {
-    it('should make HTTP request to backend', () => {
-      const user: User = {
-        email: undefined,
-        username: undefined,
-        firstName: undefined,
-        lastName: undefined,
-      };
-
-      service.getUser('1').subscribe((res) => {
-        expect(res).toEqual(user);
-      });
-
-      const req = httpTestingController.expectOne(
-        BackendConstants.END_POINT.concat('/plan/get_user_by_id/?id=1')
-      );
-      expect(req.request.method).toEqual('GET');
-      req.flush(user);
-      httpTestingController.verify();
-    });
-  });
-
   describe('getPlan', () => {
     it('should make HTTP get request to DB', () => {
       const expectedPlan: Plan = {
