+++ src/planscape/plan/urls.py
@@ -4,10 +4,9 @@ from plan.views import (create_plan, create_project, create_project_area,
                         delete, delete_projects, delete_scenarios,
                         favorite_scenario, get_plan, get_project,
                         get_project_areas, get_scenario, get_scores, 
-                        get_user_by_id, list_plans_by_owner, 
-                        list_projects_for_plan, list_scenarios_for_plan, 
-                        unfavorite_scenario, update_project, update_scenario,
-                        queue_forsys_call)
+                        list_plans_by_owner, list_projects_for_plan, 
+                        list_scenarios_for_plan, unfavorite_scenario, 
+                        update_project, update_scenario, queue_forsys_call)
 
 app_name = 'plan'
 
@@ -39,6 +38,4 @@ urlpatterns = [
     path('favorite_scenario/', favorite_scenario, name='favorite_scenario'),
     path('unfavorite_scenario/', unfavorite_scenario, name='unfavorite_scenario'),
     path('queue_forsys_call/', queue_forsys_call, name='queue_forsys_call'),
-    # Users
-    path('get_user_by_id/', get_user_by_id, name='get_user_by_id'),
 ]
