+++ src/planscape/users/serializers.py
@@ -1,5 +1,6 @@
 from dj_rest_auth.registration.serializers import RegisterSerializer
 from rest_framework import serializers
+from django.contrib.auth.models import User
 
 
 class NameRegistrationSerializer(RegisterSerializer):
@@ -11,3 +12,8 @@ class NameRegistrationSerializer(RegisterSerializer):
     user.first_name = self.validated_data.get('first_name', '')
     user.last_name = self.validated_data.get('last_name', '')
     user.save(update_fields=['first_name', 'last_name'])
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = '__all__'
