+++ src/planscape/users/urls.py
@@ -5,4 +5,5 @@ app_name = 'users'
 
 urlpatterns = [
     path('delete/', views.delete_user, name='delete'),
+    path('get_user_by_id/', views.get_user_by_id, name='get_user_by_id'),
 ]
